+++ mobile/library/java/io/envoyproxy/envoymobile/engine/EnvoyConfiguration.java
@@ -9,6 +9,7 @@ import javax.annotation.Nullable;
 
 import io.envoyproxy.envoymobile.engine.types.EnvoyHTTPFilterFactory;
 import io.envoyproxy.envoymobile.engine.types.EnvoyStringAccessor;
+import io.envoyproxy.envoymobile.engine.types.EnvoyKeyValueStore;
 
 /* Typed configuration that may be used for starting Envoy. */
 public class EnvoyConfiguration {
@@ -54,6 +55,7 @@ public class EnvoyConfiguration {
   public final String virtualClusters;
   public final List<EnvoyNativeFilterConfig> nativeFilterChain;
   public final Map<String, EnvoyStringAccessor> stringAccessors;
+  public final Map<String, EnvoyKeyValueStore> keyValueStores;
 
   private static final Pattern UNRESOLVED_KEY_PATTERN = Pattern.compile("\\{\\{ (.+) \\}\\}");
 
@@ -93,6 +95,7 @@ public class EnvoyConfiguration {
    * @param nativeFilterChain            the configuration for native filters.
    * @param httpPlatformFilterFactories  the configuration for platform filters.
    * @param stringAccessors              platform string accessors to register.
+   * @param keyValueStores               platform key-value store implementations.
    */
   public EnvoyConfiguration(
       Boolean adminInterfaceEnabled, String grpcStatsDomain, @Nullable Integer statsdPort,
@@ -107,7 +110,8 @@ public class EnvoyConfiguration {
       String appVersion, String appId, TrustChainVerification trustChainVerification,
       String virtualClusters, List<EnvoyNativeFilterConfig> nativeFilterChain,
       List<EnvoyHTTPFilterFactory> httpPlatformFilterFactories,
-      Map<String, EnvoyStringAccessor> stringAccessors) {
+      Map<String, EnvoyStringAccessor> stringAccessors,
+      Map<String, EnvoyKeyValueStore> keyValueStores) {
     this.adminInterfaceEnabled = adminInterfaceEnabled;
     this.grpcStatsDomain = grpcStatsDomain;
     this.statsdPort = statsdPort;
@@ -140,6 +144,7 @@ public class EnvoyConfiguration {
     this.nativeFilterChain = nativeFilterChain;
     this.httpPlatformFilterFactories = httpPlatformFilterFactories;
     this.stringAccessors = stringAccessors;
+    this.keyValueStores = keyValueStores;
   }
 
   /**